updating review comments
