+++ docker-compose.yml
@@ -75,7 +75,7 @@ services:
       GL_CODE: ${GL_CODE}
       ORBC_SERVICE_ACCOUNT_CLIENT_ID: ${ORBC_SERVICE_ACCOUNT_CLIENT_ID}
       ORBC_SERVICE_ACCOUNT_CLIENT_SECRET: ${ORBC_SERVICE_ACCOUNT_CLIENT_SECRET}
-      KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE: ${KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE}
+      ORBC_SERVICE_ACCOUNT_AUDIENCE : ${ORBC_SERVICE_ACCOUNT_AUDIENCE }
     healthcheck:
       test: ["CMD", "curl", "-f", "http://localhost:5000/"]
       interval: 1m30s
@@ -132,7 +132,7 @@ services:
       FRONTEND_URL: "http://frontend:3000"
       ORBC_SERVICE_ACCOUNT_CLIENT_ID: ${ORBC_SERVICE_ACCOUNT_CLIENT_ID}
       ORBC_SERVICE_ACCOUNT_CLIENT_SECRET: ${ORBC_SERVICE_ACCOUNT_CLIENT_SECRET}
-      KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE: ${KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE}
+      ORBC_SERVICE_ACCOUNT_AUDIENCE : ${ORBC_SERVICE_ACCOUNT_AUDIENCE }
     healthcheck:
       test: ["CMD", "curl", "-f", "http://localhost:5001/"]
       interval: 1m30s
