+++ dops/Dockerfile
@@ -67,7 +67,7 @@ ENV ACCESS_API_URL ${ACCESS_API_URL}
 ENV DOPS_URL ${DOPS_URL}
 ENV FRONTEND_URL ${FRONTEND_URL}
 ENV ORBC_SERVICE_ACCOUNT_CLIENT_ID ${ORBC_SERVICE_ACCOUNT_CLIENT_ID}
-ENV KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE ${KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE}
+ENV ORBC_SERVICE_ACCOUNT_AUDIENCE  ${ORBC_SERVICE_ACCOUNT_AUDIENCE }
 
 # Copy production files from build
 COPY --from=builder /app/package*.json ./
