+++ dops/src/modules/dgen/dgen.controller.ts
@@ -87,7 +87,6 @@ export class DgenController {
   }
 
   @Roles(Role.GENERATE_REPORT)
-  @UseGuards(JwtAuthGuard, JwtServiceAccountAuthGuard)
   @Post('/report/render')
   async generateReport(
     @Req() request: Request,
