+++ vehicles/Dockerfile
@@ -53,7 +53,7 @@ ENV PAYBC_BASE_URL ${PAYBC_BASE_URL}
 ENV GL_CODE ${GL_CODE}
 ENV PAYBC_REDIRECT ${PAYBC_REDIRECT}
 ENV ORBC_SERVICE_ACCOUNT_CLIENT_ID ${ORBC_SERVICE_ACCOUNT_CLIENT_ID}
-ENV KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE ${KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE}
+ENV ORBC_SERVICE_ACCOUNT_AUDIENCE  ${ORBC_SERVICE_ACCOUNT_AUDIENCE }
 ENV PERMIT_ISSUE_LIMIT ${PERMIT_ISSUE_LIMIT}
 ENV PERMIT_DOCUMENT_GENERATION_LIMIT ${PERMIT_DOCUMENT_GENERATION_LIMIT}
