+++ vehicles/src/modules/auth/jwt-service-account.strategy.ts
@@ -25,7 +25,7 @@ export class JwtServiceAccountStrategy extends PassportStrategy(
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
       ignoreExpiration:
         process.env.KEYCLOAK_IGNORE_EXP === 'true' ? true : false,
-      audience: process.env.KEYCLOAK_SERVICE_ACCOUNT_AUDIENCE,
+      audience: process.env.ORBC_SERVICE_ACCOUNT_AUDIENCE ,
       issuer: `${process.env.KEYCLOAK_ISSUER_URL}`,
       algorithms: ['RS256'],
       passReqToCallback: true,
