+++ vehicles/src/modules/permit-application-payment/application/application.controller.ts
@@ -14,7 +14,7 @@ import { ExceptionDto } from '../../../common/exception/exception.dto';
 import { ResultDto } from './dto/response/result.dto';
 import { PermitReceiptDocumentService } from '../permit-receipt-document/permit-receipt-document.service';
 import { JwtServiceAccountAuthGuard } from 'src/common/guard/jwt-sa-auth.guard';
-import { PermitIdDto } from 'src/modules/common/dto/request/permit-id.dto';
+import { PermitIdDto } from 'src/modules/permit-application-payment/permit/dto/request/permit-id.dto';
 
 @ApiBearerAuth()
 @ApiTags('Application')
@@ -53,7 +53,7 @@ export class ApplicationController {
       'Returns a list of updated application ids or throws exceptions for unauthorized access or operational failures.',
   })
   @UseGuards(JwtServiceAccountAuthGuard)
-  @Post('/scheduler/issue')
+  @Post('/issue')
   async issuePermitSchedule(
     @Req() request: Request,
     @Body() permit: PermitIdDto,
@@ -64,41 +64,5 @@ export class ApplicationController {
       permit.ids,
     );
     return result;
-  }
-
-  /**
-   * A POST method defined with the @Post() decorator and a route of /scheduler/document
-   * that generates permit and receipt document for given application ids
-   * This method only works for ORBC Service account.
-   * @param request
-   * @param PermitIdDto
-   * @returns The ids of new voided/revoked permit a in response object {@link string}
-   *
-   */
-  @ApiOperation({
-    summary: 'Generate permit and receipt document for given application ids',
-    description:
-      'Generate permit and receipt document for given application ids',
-  })
-  @UseGuards(JwtServiceAccountAuthGuard)
-  @Post('/documents')
-  async generateDocument(@Req() request: Request, @Body() permit: PermitIdDto) {
-    const currentUser = request.user as IUserJWT;
-    await this.permitReceiptDocumentService.generatePermitDocuments(
-      currentUser,
-      permit.ids,
-    );
-    return 'success';
-  }
-
-  @UseGuards(JwtServiceAccountAuthGuard)
-  @Post('/receipts')
-  async generateReceipt(@Req() request: Request, @Body() permit: PermitIdDto) {
-    const currentUser = request.user as IUserJWT;
-    await this.permitReceiptDocumentService.generateReceiptDocuments(
-      currentUser,
-      permit.ids,
-    );
-    return 'success';
   }
 }
